requirements.txt
streamlit
pandas
numpy
matplotlib
scipy
streamlit-authenticator
PyYAML